[Edge] Hardy Barth cPH1: remove EVCS, move Ecb1Handler/Parent to EVSE package
Drops the legacy EVCS component (Evcs.HardyBarth.cPH1) and its tests.
Moves Ecb1Handler and Ecb1Parent into the EVSE package; integrates the
ECB1 channel definitions into EvseChargePointHardyBarthEcb1. The EVSE
component (Evse.ChargePoint.HardyBarth.cPH1) is now self-contained.
